+++ test/manual/multipage/src/assets/scripts/common.js
@@ -1,0 +1,1 @@
+console.log('>> Common script is loaded.');